Struct google_plusdomains1::api::Person

pub struct Person {
    pub about_me: Option<String>,
    pub birthday: Option<String>,
    pub bragging_rights: Option<String>,
    pub circled_by_count: Option<i32>,
    pub cover: Option<PersonCover>,
    pub current_location: Option<String>,
    pub display_name: Option<String>,
    pub domain: Option<String>,
    pub emails: Option<Vec<PersonEmails>>,
    pub etag: Option<String>,
    pub gender: Option<String>,
    pub id: Option<String>,
    pub image: Option<PersonImage>,
    pub is_plus_user: Option<bool>,
    pub kind: Option<String>,
    pub name: Option<PersonName>,
    pub nickname: Option<String>,
    pub object_type: Option<String>,
    pub occupation: Option<String>,
    pub organizations: Option<Vec<PersonOrganizations>>,
    pub places_lived: Option<Vec<PersonPlacesLived>>,
    pub plus_one_count: Option<i32>,
    pub relationship_status: Option<String>,
    pub skills: Option<String>,
    pub tagline: Option<String>,
    pub url: Option<String>,
    pub urls: Option<Vec<PersonUrls>>,
    pub verified: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get people (response)

Fields
about_me: Option<String>

A short biography for this person.

birthday: Option<String>

The person's date of birth, represented as YYYY-MM-DD.

bragging_rights: Option<String>

The "bragging rights" line of this person.

circled_by_count: Option<i32>

For followers who are visible, the number of people who have added this person or page to a circle.

cover: Option<PersonCover>

The cover photo content.

current_location: Option<String>

(this field is not currently used)

display_name: Option<String>

The name of this person, which is suitable for display.

domain: Option<String>

The hosted domain name for the user's Google Apps account. For instance, example.com. The plus.profile.emails.read or email scope is needed to get this domain name.

emails: Option<Vec<PersonEmails>>

A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile. The plus.profile.emails.read scope is needed to retrieve these email addresses, or the email scope can be used to retrieve just the Google account email address.

etag: Option<String>

ETag of this response for caching purposes.

gender: Option<String>

The person's gender. Possible values include, but are not limited to, the following values:

"male" - Male gender.

"female" - Female gender.

"other" - Other.

id: Option<String>

The ID of this person.

image: Option<PersonImage>

The representation of the person's profile photo.

is_plus_user: Option<bool>

Whether this user has signed up for Google+.

kind: Option<String>

Identifies this resource as a person. Value: "plus#person".

name: Option<PersonName>

An object representation of the individual components of a person's name.

nickname: Option<String>

The nickname of this person.

object_type: Option<String>

Type of person within Google+. Possible values include, but are not limited to, the following values:

"person" - represents an actual person.

"page" - represents a page.

occupation: Option<String>

The occupation of this person.

organizations: Option<Vec<PersonOrganizations>>

A list of current or past organizations with which this person is associated.

places_lived: Option<Vec<PersonPlacesLived>>

A list of places where this person has lived.

plus_one_count: Option<i32>

If a Google+ Page, the number of people who have +1'd this page.

relationship_status: Option<String>

The person's relationship status. Possible values include, but are not limited to, the following values:

"single" - Person is single.

"in_a_relationship" - Person is in a relationship.

"engaged" - Person is engaged.

"married" - Person is married.

"its_complicated" - The relationship is complicated.

"open_relationship" - Person is in an open relationship.

"widowed" - Person is widowed.

"in_domestic_partnership" - Person is in a domestic partnership.

"in_civil_union" - Person is in a civil union.

skills: Option<String>

The person's skills.

tagline: Option<String>

The brief description (tagline) of this person.

url: Option<String>

The URL of this person's profile.

urls: Option<Vec<PersonUrls>>

A list of URLs for this person.

verified: Option<bool>

Whether the person or Google+ Page has been verified.