Struct google_plusdomains1::Activity

pub struct Activity {
    pub place_name: Option<String>,
    pub updated: Option<String>,
    pub object: Option<ActivityObject>,
    pub verb: Option<String>,
    pub radius: Option<String>,
    pub etag: Option<String>,
    pub address: Option<String>,
    pub annotation: Option<String>,
    pub kind: Option<String>,
    pub title: Option<String>,
    pub url: Option<String>,
    pub place_id: Option<String>,
    pub actor: Option<ActivityActor>,
    pub id: Option<String>,
    pub access: Option<Acl>,
    pub geocode: Option<String>,
    pub location: Option<Place>,
    pub provider: Option<ActivityProvider>,
    pub crosspost_source: Option<String>,
    pub published: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert activities (request|response)

get activities (response)

Fields
place_name: Option<String>

Name of the place where this activity occurred.

updated: Option<String>

The time at which this activity was last updated. Formatted as an RFC 3339 timestamp.

object: Option<ActivityObject>

The object of this activity.

verb: Option<String>

This activity's verb, which indicates the action that was performed. Possible values include, but are not limited to, the following values: - "post" - Publish content to the stream. - "share" - Reshare an activity.

radius: Option<String>

Radius, in meters, of the region where this activity occurred, centered at the latitude and longitude identified in geocode.

etag: Option<String>

ETag of this response for caching purposes.

address: Option<String>

Street address where this activity occurred.

annotation: Option<String>

Additional content added by the person who shared this activity, applicable only when resharing an activity.

kind: Option<String>

Identifies this resource as an activity. Value: "plus#activity".

title: Option<String>

Title of this activity.

url: Option<String>

The link to this activity.

place_id: Option<String>

ID of the place where this activity occurred.

actor: Option<ActivityActor>

The person who performed this activity.

id: Option<String>

The ID of this activity.

access: Option<Acl>

Identifies who has access to see this activity.

geocode: Option<String>

Latitude and longitude where this activity occurred. Format is latitude followed by longitude, space separated.

location: Option<Place>

The location where this activity occurred.

provider: Option<ActivityProvider>

The service provider that initially published this activity.

crosspost_source: Option<String>

If this activity is a crosspost from another system, this property specifies the ID of the original activity.

published: Option<String>

The time at which this activity was initially published. Formatted as an RFC 3339 timestamp.