Struct google_plusdomains1::ActivityObjectAttachments

pub struct ActivityObjectAttachments {
    pub preview_thumbnails: Option<Vec<ActivityObjectAttachmentsPreviewThumbnails>>,
    pub display_name: Option<String>,
    pub thumbnails: Option<Vec<ActivityObjectAttachmentsThumbnails>>,
    pub full_image: Option<ActivityObjectAttachmentsFullImage>,
    pub url: Option<String>,
    pub image: Option<ActivityObjectAttachmentsImage>,
    pub content: Option<String>,
    pub embed: Option<ActivityObjectAttachmentsEmbed>,
    pub id: Option<String>,
    pub object_type: Option<String>,
}

The media objects attached to this activity.

This type is not used in any activity, and only used as part of another schema.

Fields
preview_thumbnails: Option<Vec<ActivityObjectAttachmentsPreviewThumbnails>>

When previewing, these are the optional thumbnails for the post. When posting an article, choose one by setting the attachment.image.url property. If you don't choose one, one will be chosen for you.

display_name: Option<String>

The title of the attachment, such as a photo caption or an article title.

thumbnails: Option<Vec<ActivityObjectAttachmentsThumbnails>>

If the attachment is an album, this property is a list of potential additional thumbnails from the album.

full_image: Option<ActivityObjectAttachmentsFullImage>

The full image URL for photo attachments.

url: Option<String>

The link to the attachment, which should be of type text/html.

image: Option<ActivityObjectAttachmentsImage>

The preview image for photos or videos.

content: Option<String>

If the attachment is an article, this property contains a snippet of text from the article. It can also include descriptions for other types.

embed: Option<ActivityObjectAttachmentsEmbed>

If the attachment is a video, the embeddable link.

id: Option<String>

The ID of the attachment.

object_type: Option<String>

The type of media object. Possible values include, but are not limited to, the following values: - "photo" - A photo. - "album" - A photo album. - "video" - A video. - "article" - An article, specified by a link.