Struct google_plusdomains1::AudiencesFeed

pub struct AudiencesFeed {
    pub next_page_token: Option<String>,
    pub items: Option<Vec<Audience>>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub total_items: Option<i32>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list audiences (response)

Fields
next_page_token: Option<String>

The continuation token, which is used to page through large result sets. Provide this value in a subsequent request to return the next page of results.

items: Option<Vec<Audience>>

The audiences in this result.

kind: Option<String>

Identifies this resource as a collection of audiences. Value: "plus#audienceFeed".

etag: Option<String>

ETag of this response for caching purposes.

total_items: Option<i32>

The total number of ACL entries. The number of entries in this response may be smaller due to paging.