Struct google_plusdomains1::CircleMethods

pub struct CircleMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on circle resources. It is not used directly, but through the PlusDomains hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_plusdomains1 as plusdomains1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plusdomains1::PlusDomains;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = PlusDomains::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_people(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `remove(...)`, `remove_people(...)` and `update(...)`
// to build up your call.
let rb = hub.circles();

Methods

impl<'a, C, A> CircleMethods<'a, C, A>

fn remove_people(&self, circle_id: &str) -> CircleRemovePeopleCall<'a, C, A>

Create a builder to help you perform the following task:

Remove a person from a circle.

Arguments

circleId - The ID of the circle to remove the person from.

fn insert(&self, request: Circle, user_id: &str) -> CircleInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a new circle for the authenticated user.

Arguments

request - No description provided.

userId - The ID of the user to create the circle on behalf of. The value "me" can be used to indicate the authenticated user.

fn add_people(&self, circle_id: &str) -> CircleAddPeopleCall<'a, C, A>

Create a builder to help you perform the following task:

Add a person to a circle. Google+ limits certain circle operations, including the number of circle adds. Learn More.

Arguments

circleId - The ID of the circle to add the person to.

fn get(&self, circle_id: &str) -> CircleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a circle.

Arguments

circleId - The ID of the circle to get.

fn list(&self, user_id: &str) -> CircleListCall<'a, C, A>

Create a builder to help you perform the following task:

List all of the circles for a user.

Arguments

userId - The ID of the user to get circles for. The special value "me" can be used to indicate the authenticated user.

fn update(&self, request: Circle, circle_id: &str) -> CircleUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update a circle's description.

Arguments

request - No description provided.

circleId - The ID of the circle to update.

fn patch(&self, request: Circle, circle_id: &str) -> CirclePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update a circle's description. This method supports patch semantics.

Arguments

request - No description provided.

circleId - The ID of the circle to update.

fn remove(&self, circle_id: &str) -> CircleRemoveCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a circle.

Arguments

circleId - The ID of the circle to delete.