Struct google_plusdomains1::PersonName

pub struct PersonName {
    pub honorific_prefix: Option<String>,
    pub given_name: Option<String>,
    pub middle_name: Option<String>,
    pub family_name: Option<String>,
    pub formatted: Option<String>,
    pub honorific_suffix: Option<String>,
}

An object representation of the individual components of a person's name.

This type is not used in any activity, and only used as part of another schema.

Fields
honorific_prefix: Option<String>

The honorific prefixes (such as "Dr." or "Mrs.") for this person.

given_name: Option<String>

The given name (first name) of this person.

middle_name: Option<String>

The middle name of this person.

family_name: Option<String>

The family name (last name) of this person.

formatted: Option<String>

The full name of this person, including middle names, suffixes, etc.

honorific_suffix: Option<String>

The honorific suffixes (such as "Jr.") for this person.