Struct google_plusdomains1::Media

pub struct Media {
    pub updated: Option<String>,
    pub media_created_time: Option<String>,
    pub video_duration: Option<String>,
    pub size_bytes: Option<String>,
    pub height: Option<u32>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub exif: Option<MediaExif>,
    pub author: Option<MediaAuthor>,
    pub url: Option<String>,
    pub media_url: Option<String>,
    pub video_status: Option<String>,
    pub summary: Option<String>,
    pub width: Option<u32>,
    pub etag: Option<String>,
    pub streams: Option<Vec<Videostream>>,
    pub published: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert media (request|response)

Fields

updated

The time at which this media was last updated. This includes changes to media metadata. Formatted as an RFC 3339 timestamp.

media_created_time

The time at which this media was originally created in UTC. Formatted as an RFC 3339 timestamp that matches this example: 2010-11-25T14:30:27.655Z

video_duration

The duration in milliseconds of this video.

size_bytes

The size in bytes of this video.

height

The height in pixels of the original image.

id

ID of this media, which is generated by the API.

kind

The type of resource.

display_name

The display name for this media.

exif

Exif information of the media item.

author

The person who uploaded this media.

url

The URL for the page that hosts this media.

media_url

The URL of this photo or video's still image.

video_status

The encoding status of this video. Possible values are: - "UPLOADING" - Not all the video bytes have been received. - "PENDING" - Video not yet processed. - "FAILED" - Video processing failed. - "READY" - A single video stream is playable. - "FINAL" - All video streams are playable.

summary

A description, or caption, for this media.

width

The width in pixels of the original image.

etag

ETag of this response for caching purposes.

streams

The list of video streams for this video. There might be several different streams available for a single video, either Flash or MPEG, of various sizes

published

The time at which this media was uploaded. Formatted as an RFC 3339 timestamp.