Struct google_plusdomains1::Person

pub struct Person {
    pub bragging_rights: Option<String>,
    pub relationship_status: Option<String>,
    pub image: Option<PersonImage>,
    pub domain: Option<String>,
    pub places_lived: Option<Vec<PersonPlacesLived>>,
    pub birthday: Option<String>,
    pub display_name: Option<String>,
    pub nickname: Option<String>,
    pub emails: Option<Vec<PersonEmails>>,
    pub occupation: Option<String>,
    pub organizations: Option<Vec<PersonOrganizations>>,
    pub kind: Option<String>,
    pub verified: Option<bool>,
    pub name: Option<PersonName>,
    pub is_plus_user: Option<bool>,
    pub skills: Option<String>,
    pub tagline: Option<String>,
    pub current_location: Option<String>,
    pub cover: Option<PersonCover>,
    pub id: Option<String>,
    pub url: Option<String>,
    pub etag: Option<String>,
    pub plus_one_count: Option<i32>,
    pub urls: Option<Vec<PersonUrls>>,
    pub gender: Option<String>,
    pub circled_by_count: Option<i32>,
    pub object_type: Option<String>,
    pub about_me: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get people (response)

Fields

bragging_rights

The "bragging rights" line of this person.

relationship_status

The person's relationship status. Possible values include, but are not limited to, the following values: - "single" - Person is single. - "in_a_relationship" - Person is in a relationship. - "engaged" - Person is engaged. - "married" - Person is married. - "its_complicated" - The relationship is complicated. - "open_relationship" - Person is in an open relationship. - "widowed" - Person is widowed. - "in_domestic_partnership" - Person is in a domestic partnership. - "in_civil_union" - Person is in a civil union.

image

The representation of the person's profile photo.

domain

The hosted domain name for the user's Google Apps account. For instance, example.com. The plus.profile.emails.read or email scope is needed to get this domain name.

places_lived

A list of places where this person has lived.

birthday

The person's date of birth, represented as YYYY-MM-DD.

display_name

The name of this person, which is suitable for display.

nickname

The nickname of this person.

emails

A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile. The plus.profile.emails.read scope is needed to retrieve these email addresses, or the email scope can be used to retrieve just the Google account email address.

occupation

The occupation of this person.

organizations

A list of current or past organizations with which this person is associated.

kind

Identifies this resource as a person. Value: "plus#person".

verified

Whether the person or Google+ Page has been verified.

name

An object representation of the individual components of a person's name.

is_plus_user

Whether this user has signed up for Google+.

skills

The person's skills.

tagline

The brief description (tagline) of this person.

current_location

(this field is not currently used)

cover

The cover photo content.

id

The ID of this person.

url

The URL of this person's profile.

etag

ETag of this response for caching purposes.

plus_one_count

If a Google+ Page, the number of people who have +1'd this page.

urls

A list of URLs for this person.

gender

The person's gender. Possible values include, but are not limited to, the following values: - "male" - Male gender. - "female" - Female gender. - "other" - Other.

circled_by_count

For followers who are visible, the number of people who have added this person or page to a circle.

object_type

Type of person within Google+. Possible values include, but are not limited to, the following values: - "person" - represents an actual person. - "page" - represents a page.

about_me

A short biography for this person.